+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.2.0] - 2025-08-07
+
+### Added
+
+- Arbitrary filters now available on all search endpoints
+- Search combines full-text and semantic using reciprocal rank fusion
+
+### Changed
+
+- Search has 10 results by default, max 100 results
+
 ## [2.1.2] - 2025-07-30
 
 ### Fixed